requirements.txt
fastapi==0.105.0
uvicorn==0.24.0.post1
google-cloud-datastore==2.19.0
